refactor: 관심사에 따라 함수 분리
